# stormborne-booking-backend/docs/swagger.yaml
basePath: /
definitions:
  controller.forgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  controller.resetPasswordRequest:
    properties:
      password:
        minLength: 6
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
  error.AppError:
    properties:
      code:
        type: string
      error: {}
      message:
        type: string
    type: object
  model.AvatarType:
    enum:
    - predefined
    - uploaded
    type: string
    x-enum-varnames:
    - PREDEFINED
    - UPLOADED
  model.Customer:
    properties:
      age:
        type: integer
      avatarType:
        $ref: '#/definitions/model.AvatarType'
      avatarURL:
        type: string
      countryCode:
        type: string
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        $ref: '#/definitions/model.Gender'
      id:
        type: integer
      isEmailVerified:
        type: boolean
      isPhoneVerified:
        type: boolean
      lastName:
        type: string
      password:
        type: string
      passwordResetToken:
        type: string
      passwordTokenExpiryTime:
        type: string
      phoneNumber:
        type: string
      updatedAt:
        type: string
      userRole:
        type: string
      username:
        type: string
    type: object
  model.Gender:
    enum:
    - male
    - female
    - prefer_not_to
    type: string
    x-enum-varnames:
    - MALE
    - FEMALE
    - PREFER_NOT_TO_SAY
...
  request.CaptchaVerifyRequest:
    properties:
      captcha_id:
        type: string
      slider_position_x:
        type: number
    required:
    - captcha_id
    - slider_position_x
    type: object
  request.CustomerLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    type: object
  request.CustomerSignupRequest:
    properties:
      age:
        minimum: 1
        type: integer
      captcha_id:
        type: string
      countryCode:
        type: string
      email:
        type: string
      firstName:
        maxLength: 100
        minLength: 2
        type: string
      gender:
        $ref: '#/definitions/model.Gender'
      lastName:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
      phoneNumber:
        type: string
      slider_position_x:
        type: number
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - age
    - captcha_id
    - countryCode
    - email
    - firstName
    - gender
    - lastName
    - password
    - phoneNumber
    - slider_position_x
    - username
    type: object
  request.CustomerUpdateProfileRequest:
    properties:
      age:
        minimum: 1
        type: integer
      email:
        type: string
      first_name:
        maxLength: 20
        minLength: 2
        type: string
      last_name:
        maxLength: 20
        minLength: 2
        type: string
    type: object
...
  response.CaptchaGenerateResponse:
    properties:
      captcha_id:
        type: string
      master_image:
        type: string
      tile_height:
        type: integer
      tile_image:
        type: string
      tile_width:
        type: integer
      tile_x:
        type: integer
      tile_y:
        type: integer
    type: object
  response.LoginResponse:
    properties:
      age:
        type: integer
      avatar_type:
        $ref: '#/definitions/model.AvatarType'
      avatar_url:
        type: string
      gender:
        $ref: '#/definitions/model.Gender'
      id:
        type: string
      role:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a skyfox
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Booking API
  version: "1.0"
paths:
...
  /api/captcha/generate:
    get:
      description: 'Returns a new slide captcha: background image, tile image and
        initial positions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CaptchaGenerateResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Generate slide captcha
      tags:
      - captcha
  /api/captcha/verify:
    post:
      consumes:
      - application/json
      description: Verifies the user's slider position against the stored target X
      parameters:
      - description: Captcha verify payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CaptchaVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify slide captcha
      tags:
      - captcha
...
  /customer/profile:
    get:
      consumes:
      - application/json
      description: Returns the profile details of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Profile fetched successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Failed to fetch profile
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Invalid user ID or server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: View Profile
      tags:
      - OnlineCustomer
  /customer/signup:
    post:
      consumes:
      - application/json
      description: Register new online customer
      parameters:
      - description: Signup Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CustomerSignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      summary: Customer Signup
      tags:
      - OnlineCustomer
  /forgot-password:
    post:
      consumes:
      - application/json
      description: Validates email and sends a password reset link to the user
      parameters:
      - description: Email of the user
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.forgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email Found in DB
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or Forgot password failed
          schema:
            additionalProperties: true
            type: object
      summary: Forgot Password
      tags:
      - Authentication
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate online customer
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CustomerLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      summary: Customer Login
      tags:
      - OnlineCustomer
...
  /profile:
    put:
      consumes:
      - application/json
      description: Updates editable profile fields (first name, last name, email)
        of the authenticated user.
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update Profile Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CustomerUpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or validation error
          schema:
            $ref: '#/definitions/error.AppError'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Email already in use
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Profile
      tags:
      - OnlineCustomer
  /reset-password:
    get:
      consumes:
      - application/json
      description: Retrieves the reset token from the URL query parameters to initiate
        the password reset process
      parameters:
      - description: Reset Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token received successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Token missing or invalid
          schema:
            additionalProperties: true
            type: object
      summary: View Reset Password Page
      tags:
      - Authentication
    post:
      consumes:
      - application/json
      description: Verifies the reset token and updates the password in the database
      parameters:
      - description: Token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.resetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or Reset password failed
          schema:
            additionalProperties: true
            type: object
      summary: Reset Password
      tags:
      - Authentication
...
  /verify-reset-token:
    get:
      consumes:
      - application/json
      description: Validates the password reset token before allowing the user to
        reset the password
      parameters:
      - description: Password reset token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Valid token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Token missing or invalid
          schema:
            additionalProperties: true
            type: object
      summary: Verify Reset Token
      tags:
      - Authentication
swagger: "2.0"